# specs/005-api-webhooks/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EmailKit Public API
  description: |
    EmailKit API for email verification, webhook management, and credit balance queries.

    ## Authentication
    All endpoints (except `/health`) require Bearer token authentication using an API key.
    ```
    Authorization: Bearer ek_YOUR_API_KEY
    ```

    ## Rate Limiting
    Rate limits are enforced per-user based on subscription tier. All responses include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when window resets

    ## Idempotency
    POST requests require an `Idempotency-Key` header (UUID v4). Duplicate requests
    within 24 hours return the cached response.

    ## Test Mode
    API keys with `ek_test_` prefix return mock responses without consuming credits.
  version: '1.0.0'
  contact:
    name: EmailKit Support
    email: support@emailkit.io
servers:
  - url: https://api.emailkit.io/api/v1
    description: Production API
security:
  - BearerAuth: []
...
paths:
...
  /verify:
    post:
      operationId: verifySingle
      tags: [Verification]
      summary: Verify single email
      description: Verify a single email address. Consumes 1 credit.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  description: Email address to verify
      responses:
        '200':
          description: Verification result
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /verify/batch:
    post:
      operationId: verifyBatch
      tags: [Verification]
      summary: Verify batch of emails
      description: |
        Verify up to 100 emails synchronously. Consumes 1 credit per email.
        Returns an array of results in the same order as input.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emails]
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                  minItems: 1
                  maxItems: 100
                  description: Array of email addresses to verify
      responses:
        '200':
          description: Batch verification results
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerificationResult'
                  creditsUsed:
                    type: integer
                    description: Total credits consumed
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /verify/bulk:
    post:
      operationId: createBulkJob
      tags: [Verification]
      summary: Create bulk verification job
      description: |
        Submit a bulk verification job for asynchronous processing.
        Accepts CSV file or JSON array. Job limits are tier-based:
        - Starter: 10,000 emails
        - Growth: 25,000 emails
        - Pro: 50,000 emails
        - Scale: 100,000 emails
        - Titan: 500,000 emails
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with email addresses
          application/json:
            schema:
              type: object
              required: [emails]
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    format: email
                  description: Array of email addresses
      responses:
        '202':
          description: Bulk job created
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '409':
          description: Concurrent job limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Job size exceeds tier limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
  /verify/bulk/{jobId}:
    get:
      operationId: getBulkJobStatus
      tags: [Verification]
      summary: Get bulk job status
      description: Retrieve the current status and progress of a bulk verification job.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /verify/bulk/{jobId}/results:
    get:
      operationId: getBulkJobResults
      tags: [Verification]
      summary: Download bulk job results
      description: |
        Download the results of a completed bulk verification job as CSV.
        Results expire after 14 days.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CSV file with results
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="results.csv"'
          content:
            text/csv:
              schema:
                type: string
        '302':
          description: Redirect to download URL
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Results expired (14+ days old)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Use your API key as the bearer token.
        ```
        Authorization: Bearer ek_YOUR_API_KEY
        ```
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: UUID v4 for request deduplication. Required for all POST requests.
  headers:
    X-Request-ID:
      schema:
        type: string
      description: Unique request identifier for debugging
    X-RateLimit-Limit:
      schema:
        type: integer
      description: Maximum requests per window
    X-RateLimit-Remaining:
      schema:
        type: integer
      description: Remaining requests in current window
    X-RateLimit-Reset:
      schema:
        type: integer
      description: Unix timestamp when rate limit window resets
  schemas:
    VerificationResult:
      type: object
      properties:
        email:
          type: string
          format: email
        status:
          type: string
          enum: [valid, invalid, risky, unknown]
        score:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score (0.0 to 1.0)
        deliverability:
          type: string
          enum: [deliverable, undeliverable, risky, unknown]
        attributes:
          type: object
          properties:
            disposable:
              type: boolean
            freeProvider:
              type: boolean
            roleAccount:
              type: boolean
            catchAll:
              type: boolean
            mxRecordsFound:
              type: boolean
            smtpValid:
              type: boolean
        requestId:
          type: string
    BulkJob:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        totalCount:
          type: integer
        processedCount:
          type: integer
        progress:
          type: number
          description: Percentage complete (0-100)
        results:
          type: object
          nullable: true
          properties:
            valid:
              type: integer
            invalid:
              type: integer
            risky:
              type: integer
            unknown:
              type: integer
        resultUrl:
          type: string
          format: uri
          nullable: true
          description: URL to download results (available when completed)
        resultExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        requestId:
          type: string
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details
        requestId:
          type: string
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Invalid or expired API key
    Forbidden:
      description: Access denied to this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
            message: You do not have access to this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: The requested resource was not found
    InsufficientCredits:
      description: Insufficient credits to perform operation
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  creditsRequired:
                    type: integer
                  creditsAvailable:
                    type: integer
          example:
            error: insufficient_credits
            message: You need 100 credits but only have 50
            creditsRequired: 100
            creditsAvailable: 50
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  fields:
                    type: object
                    additionalProperties:
                      type: string
          example:
            error: validation_error
            message: Request validation failed
            fields:
              email: Invalid email format
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Too many requests. Please retry after 5 seconds.